+++ collect_app/src/main/java/org/odk/collect/android/tasks/AuditEventSaveTask.java
@@ -4,9 +4,7 @@ package org.odk.collect.android.tasks;
 import android.os.AsyncTask;
 import androidx.annotation.NonNull;
 
-import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.logic.AuditEvent;
-import org.odk.collect.android.logic.FormController;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -48,10 +46,8 @@ public class AuditEventSaveTask extends AsyncTask<AuditEvent, Void, Void> {
             }
             if (params.length > 0) {
                 for (AuditEvent aev : params) {
-                    if (shouldEventBeLogged(aev)) {
-                        fw.write(aev.toString() + "\n");
-                        Timber.i("Log audit event: %s", aev.toString());
-                    }
+                    fw.write(aev.toString() + "\n");
+                    Timber.i("Log audit event: %s", aev.toString());
                 }
             }
         } catch (IOException e) {
@@ -68,19 +64,6 @@ public class AuditEventSaveTask extends AsyncTask<AuditEvent, Void, Void> {
             }
         }
         return null;
-    }
-
-    /*
-     Question which is in field-list group should be logged only if tracking changes option is
-     enabled and its answer has changed
-      */
-    private boolean shouldEventBeLogged(AuditEvent aev) {
-        FormController formController = Collect.getInstance().getFormController();
-        if (aev.getAuditEventType().equals(AuditEvent.AuditEventType.QUESTION) && formController != null) {
-            return !formController.indexIsInFieldList(aev.getFormIndex())
-                    || (aev.newAnswerAppeared() && isTrackingChangesEnabled);
-        }
-        return true;
     }
 
     private boolean updateHeaderIfNeeded() {
